+++ internal/acceptance_tests/blocks/logging_logshuttle_nested.tf
@@ -1,0 +1,7 @@
+logging_logshuttle {
+  name = "{{.LOGGING_LOGSHUTTLE_NAME}}"
+  url  = "https://east.logplex.io/logs"
+  authentication = {
+    token = "test-logshuttle-token"
+  }
+}