+++ internal/acceptance_tests/blocks/logging_logshuttle_nested_placement_none.tf
@@ -1,0 +1,8 @@
+logging_logshuttle {
+  name = "{{.LOGGING_LOGSHUTTLE_NAME}}"
+  url  = "https://east.logplex.io/logs"
+  authentication = {
+    token = "test-logshuttle-token"
+  }
+  placement = "none"
+}