+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - feat(kafka_logging): add support for Kafka Logging ([#111](https://github.com/fastly/terraform-provider-fastly-beta/pull/111))
 - feat(loggly_logging): add support for Loggly Logging ([#113](https://github.com/fastly/terraform-provider-fastly-beta/pull/113))
 - feat(honeycomb_logging): add support for Honeycomb Logging ([#112](https://github.com/fastly/terraform-provider-fastly-beta/pull/112))
+- feat(logshuttle_logging): add support for Log Shuttle Logging ([#117](https://github.com/fastly/terraform-provider-fastly-beta/pull/117))
 
 ### BUG FIXES: